Refactor dataset creation in DropletTrack and Emulsion classes to specify dtype as "f4" and remove pytest from module dependencies

+++ droplets/droplet_tracks.py
@@ -335,7 +335,7 @@ class DropletTrack:
 
         else:
             # create empty dataset to indicate empty emulsion
-            dataset = hdf_path.create_dataset(key, shape=())
+            dataset = hdf_path.create_dataset(key, shape=(), dtype="f4")
             dataset.attrs["droplet_class"] = "None"
 
         return dataset
